+++ src/system/drd_system_daemon.h
@@ -19,5 +19,7 @@ DrdSystemDaemon *drd_system_daemon_new(DrdConfig *config,
 gboolean drd_system_daemon_start(DrdSystemDaemon *self, GError **error);
 void drd_system_daemon_stop(DrdSystemDaemon *self);
 gboolean drd_system_daemon_set_main_loop(DrdSystemDaemon *self, GMainLoop *loop);
+guint drd_system_daemon_get_pending_client_count(DrdSystemDaemon *self);
+guint drd_system_daemon_get_remote_client_count(DrdSystemDaemon *self);
 
 G_END_DECLS