+++ src/utils/drd_frame_queue.h
@@ -4,6 +4,8 @@
 
 #include "utils/drd_frame.h"
 
+#define DRD_FRAME_QUEUE_MAX_FRAMES 3
+
 G_BEGIN_DECLS
 
 #define DRD_TYPE_FRAME_QUEUE (drd_frame_queue_get_type())
@@ -17,5 +19,6 @@ gboolean drd_frame_queue_wait(DrdFrameQueue *self,
                                gint64 timeout_us,
                                DrdFrame **out_frame);
 void drd_frame_queue_stop(DrdFrameQueue *self);
+guint64 drd_frame_queue_get_dropped_frames(DrdFrameQueue *self);
 
 G_END_DECLS